creation nav folder and move NavLink
